backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
tensorflow-cpu==2.15.0
